Add files via upload
This is Chrome extention that works to block some one whom you dont want to read about or see his name over the internet. this is just a fund project that i created and there is no commercial or any other use for it. simply for fun. i am not a developer but i am trying to learn and sharpen some skilss with programing on get hub.